ChatOpenAI_streamlit/requirements.txt
streamlit==1.29.0
langchain==0.0.350
python_dotenv==1.0.0
openai==1.4.0